+++ CHANGELOG.md
@@ -32,6 +32,7 @@ Breaking changes:
 Other notable changes:
 * `compy`:
   * New method `BaseComponent._prot_addChild()`, to simplify adding children.
+  * New method `BaseComponent.whenStopped()`.
   * New abstract class `BaseAggregateComponent`, for components that _publicly_
     allow children to be added.
 * `webapp-core`: